Put generic view functions in BaseViewDao
